refactor(RecentRecordResDto, RecentReviewListResDto, RecordController, RecordListResDto, RecordService, RecordServiceImpl, ReviewListResDto): 클래스가 아닌 생성자에 빌더패턴 적용
