+++ src/main/java/com/bookmile/backend/domain/record/dto/res/RecentRecordResDto.java
@@ -1,14 +1,11 @@
-package com.bookmile.backend.domain.record.dto.req;
+package com.bookmile.backend.domain.record.dto.res;
 
 import com.bookmile.backend.domain.record.entity.Record;
 import com.bookmile.backend.domain.user.entity.User;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-@Builder
-@AllArgsConstructor
 public class RecentRecordResDto {
     private Long userId;
     private String userImageUrl;
@@ -16,6 +13,17 @@ public class RecentRecordResDto {
     private Integer currentPage;
     private String text;
     private String recordImageUrl;
+
+    @Builder
+    public RecentRecordResDto(Long userId, String userImageUrl, String nickName, Integer currentPage, String text,
+                              String recordImageUrl) {
+        this.userId = userId;
+        this.userImageUrl = userImageUrl;
+        this.nickName = nickName;
+        this.currentPage = currentPage;
+        this.text = text;
+        this.recordImageUrl = recordImageUrl;
+    }
 
     public static RecentRecordResDto createRecentRecord(User user, Record record, String recordImage) {
         return RecentRecordResDto.builder()
