+++ src/main/java/com/bookmile/backend/domain/record/dto/res/RecordListResDto.java
@@ -3,19 +3,26 @@ package com.bookmile.backend.domain.record.dto.res;
 import com.bookmile.backend.domain.record.entity.Record;
 import java.time.LocalDateTime;
 import java.util.List;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-@Builder
-@AllArgsConstructor
 public class RecordListResDto {
     private Long recordId;
     private String text;
     private Integer currentPage;
     private LocalDateTime createdAt;
     private List<String> imageUrls;
+
+    @Builder
+    public RecordListResDto(Long recordId, String text, Integer currentPage, LocalDateTime createdAt,
+                            List<String> imageUrls) {
+        this.recordId = recordId;
+        this.text = text;
+        this.currentPage = currentPage;
+        this.createdAt = createdAt;
+        this.imageUrls = imageUrls;
+    }
 
     public static RecordListResDto createRecord(Record record, List<String> imageUrls) {
         return RecordListResDto.builder()
