+++ src/main/java/com/bookmile/backend/domain/record/service/Impl/RecordServiceImpl.java
@@ -8,9 +8,9 @@ import static com.bookmile.backend.global.common.StatusCode.USER_NOT_FOUND;
 import com.bookmile.backend.domain.group.entity.Group;
 import com.bookmile.backend.domain.image.repository.ImageRepository;
 import com.bookmile.backend.domain.image.service.Impl.ImageServiceImpl;
-import com.bookmile.backend.domain.record.dto.req.RecentRecordResDto;
 import com.bookmile.backend.domain.record.dto.req.RecordReqDto;
 import com.bookmile.backend.domain.record.dto.req.UpdateRecordReqDto;
+import com.bookmile.backend.domain.record.dto.res.RecentRecordResDto;
 import com.bookmile.backend.domain.record.dto.res.RecordListResDto;
 import com.bookmile.backend.domain.record.entity.Record;
 import com.bookmile.backend.domain.record.repository.RecordRepository;
